fixed the visualizer bug, now steadily runs visualization without output crashing, tested with headphones, needs further test with speakers

+++ Audio_Visualizer/fft.h
@@ -4,4 +4,4 @@
 #include <vector>
 #include <cmath>
 
-void fft(std::vector<float> input, float output[], size_t n); 
+void fft(std::vector<float> input, float output[], int n);
